requirements.txt
flask==3.0.2
langchain-community==0.0.27
youtube-transcript-api==0.6.2
pypdf2==3.0.1
reportlab==4.1.0
google-api-python-client==2.118.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
flask-sqlalchemy==3.1.1 